+++ src/lib.rs
@@ -17,6 +17,11 @@
 //! and shrinking is defined on a per-value basis instead of per-type, which
 //! makes it much more flexible and simplifies composition.
 //!
+//! If you have dependencies which provide QuickCheck `Arbitrary`
+//! implementations, see also the related
+//! [`proptest-quickcheck-interop`](https://crates.io/crates/proptest-quickcheck-interop)
+//! crates which enables reusing those implementations with proptest.
+//!
 //! <!-- NOREADME
 //! ## Status of this crate
 //!
@@ -87,7 +92,7 @@
 //!
 //! ```toml
 //! [dev-dependencies]
-//! proptest = "0.3.2"
+//! proptest = "0.3.4"
 //! ```
 //!
 //! and at the top of `main.rs` or `lib.rs`:
@@ -128,8 +133,19 @@
 //! '
 //! ```
 //!
-//! The first thing we should do is copy the failing case to a traditional unit
-//! test since it has exposed a bug.
+//! If we look at the top directory after the test fails, we'll see a new
+//! `proptest-regressions` directory, which contains some files corresponding
+//! to source files containing failing test cases. These are [_failure
+//! persistence_](#failure-persistence) files. The first thing we should do is
+//! add these to source control.
+//!
+//! ```text
+//! $ git add proptest-regressions
+//! ```
+//!
+//! The next thing we should do is copy the failing case to a traditional unit
+//! test since it has exposed a bug not similar to what we've tested in the
+//! past.
 //!
 //! ```rust,ignore
 //! #[test]
@@ -146,8 +162,8 @@
 //! check that the string is ASCII and reject anything that isn't.
 //!
 //! ```rust,no_run
-//! use std::ascii::AsciiExt;
-//!
+//! # use std::ascii::AsciiExt; //NOREADME
+//! # // NOREADME
 //! fn parse_date(s: &str) -> Option<(u32, u32, u32)> {
 //!     if 10 != s.len() { return None; }
 //!
@@ -278,7 +294,12 @@
 //! the `0000-06-20` and `0000-09-20` test cases _passed_.
 //!
 //! In the end, we get the date `0000-10-01`, which apparently gets parsed as
-//! `0000-00-01`. Again, let's add this as its own unit test:
+//! `0000-00-01`. Again, this failing case was added to the failure persistence
+//! file, and we should add this as its own unit test:
+//!
+//! ```text
+//! $ git add proptest-regressions
+//! ```
 //!
 //! ```rust,ignore
 //! #[test]
@@ -395,6 +416,47 @@
 //! which actually produces useful inputs. A strategy of `.{1,4096}` may be
 //! great to fuzz a C parser, but is highly unlikely to produce anything that
 //! makes it to a code generator.
+//!
+//! ## Failure Persistence
+//!
+//! By default, when Proptest finds a failing test case, it _persists_ that
+//! failing case in a file named after the source containing the failing test,
+//! but in a separate directory tree rooted at `proptest-regressions`† . Later
+//! runs of tests will replay those test cases before generating novel cases.
+//! This ensures that the test will not fail on one run and then spuriously
+//! pass on the next, and also exposes similar tests to the same
+//! known-problematic input.
+//!
+//! (†  If you do not have an obvious source directory, you may instead find
+//! files next to the source files, with a different extension.)
+//!
+//! It is recommended to check these files in to your source control so that
+//! other test runners (e.g., collaborators or a CI system) also replay these
+//! cases.
+//!
+//! Note that, by default, all tests in the same crate will share that one
+//! persistence file. If you have a very large number of tests, it may be
+//! desirable to separate them into smaller groups so the number of extra test
+//! cases that get run is reduced. This can be done by adjusting the
+//! `failure_persistence` flag on `Config`.
+//!
+//! There are two ways this persistence could theoretically be done.
+//!
+//! The immediately obvious option is to persist a representation of the value
+//! itself, for example by using Serde. While this has some advantages,
+//! particularly being resistant to changes like tweaking the input strategy,
+//! it also has a lot of problems. Most importantly, there is no way to
+//! determine whether any given value is actually within the domain of the
+//! strategy that produces it. Thus, some (likely extremely fragile) mechanism
+//! to ensure that the strategy that produced the value exactly matches the one
+//! in use in a test case would be required.
+//!
+//! The other option is to store the _seed_ that was used to produce the
+//! failing test case. This approach requires no support from the strategy or
+//! the produced value. If the strategy in use differs from the one used to
+//! produce failing case that was persisted, the seed may or may not produce
+//! the problematic value, but nonetheless produces a valid value. Due to these
+//! advantages, this is the approach Proptest uses.
 //!
 //! <!-- ENDREADME -->
 //!
@@ -648,7 +710,8 @@
 //! ```rust
 //! extern crate proptest;
 //!
-//! use proptest::test_runner::{TestError, TestRunner};
+//! use proptest::test_runner::{Config, FailurePersistence,
+//!                             TestError, TestRunner};
 //!
 //! fn some_function(v: i32) {
 //!     // Do a bunch of stuff, but crash if v > 500.
@@ -660,7 +723,11 @@
 //! // We know the function is broken, so use a purpose-built main function to
 //! // find the breaking point.
 //! fn main() {
-//!     let mut runner = TestRunner::default();
+//!     let mut runner = TestRunner::new(Config {
+//!         // Turn failure persistence off for demonstration
+//!         failure_persistence: FailurePersistence::Off,
+//!         .. Config::default()
+//!     });
 //!     let result = runner.run(&(0..10000i32), |&v| {
 //!         some_function(v);
 //!         Ok(())
@@ -1356,6 +1423,8 @@
 //! perusing the module tree below.
 
 #![deny(missing_docs)]
+
+#![cfg_attr(feature = "unstable", feature(i128_type))]
 
 extern crate bit_set;
 #[macro_use] extern crate lazy_static;